chore: extract Pydantic request schemas from restful_api.py into xinference/api/schemas/
- Move 14 request model classes to xinference/api/schemas/requests.py
- Add xinference/api/schemas/__init__.py with re-exports
- Update restful_api.py to import from .schemas (no behavior change)
- Remove unused imports (Dict, CreateCompletion) from restful_api.py

Phase 0 of API refactor: file reorganization only, no behavior change.
